Separated command line tool from UI stuff in App.
