+++ src/main/java/de/bayern/gdi/App.java
@@ -17,97 +17,74 @@
  */
 package de.bayern.gdi;
 
-import de.bayern.gdi.experimental.WFSTwoServiceHandler;
 import de.bayern.gdi.gui.Start;
-import de.bayern.gdi.services.WebService;
-import de.bayern.gdi.utils.ServiceChecker;
-
-import java.util.logging.Level;
-import java.util.logging.Logger;
-import org.apache.commons.codec.binary.Base64;
 
 /**
  * @author Sascha L. Teichmann (sascha.teichmann@intevation.de)
  */
 public class App {
 
-    private static final Logger log = Logger.getLogger(App.class.getName());
-
-    private static final String DEMO_URL =
-        "http://geoserv.weichand.de:8080/geoserver/wfs?"
-        + "service=WFS&acceptversions=2.0.0&request=GetCapabilities";
-    //TODO: Remove DEMO_URL and read Infos from args or external file
-
     private App() {
     }
 
     private static boolean runHeadless(String[] args) {
         for (String arg: args) {
-            if (arg.equals("-headless")) {
+            if (arg.equals("-headless")
+            || arg.equals("--headless")
+            || arg.equals("-h")) {
                 return true;
             }
         }
         return false;
     }
 
-    private static String getBase64EncAuth(String username, String password) {
-        if (username == null || password == null) {
-            return null;
-        } else {
-            String authString = username + ":" + password;
-            System.out.println("auth string: " + authString);
-            byte[] authEncBytes = Base64.encodeBase64(authString.getBytes());
-            String authStringEnc = new String(authEncBytes);
-            return authStringEnc;
+    private static boolean help(String[] args) {
+        for (String arg: args) {
+            if (arg.equals("-help")
+            || arg.equals("--help")
+            || arg.equals("-?")) {
+                return true;
+            }
         }
+        return false;
     }
-    private static WebService.Type serviceType(String url, String userName,
-                                               String password) {
-        return ServiceChecker.checkService(url, getBase64EncAuth(userName,
-                password));
+
+    private static void helpAndExit() {
+        System.out.println("java -jar downloader.jar [options]");
+        System.out.println("with options:");
+        System.out.println(
+            "  -?|--help|-help: Print this message and exit.");
+        System.out.println(
+            "  -h|--headless|-headless: Start command line tool.");
+        System.out.println("without options:");
+        System.out.println("  Start as GUI application.");
+        System.exit(0);
     }
 
     /**
      * @param args the command line arguments
      */
     public static void main(String[] args) {
-        if (runHeadless(args)) {
-            log.info("Running in headless mode");
-            //There could be a Service Handler for each Service that deals
-            //with command line arguments and/or a stored XML File
-            String userName = null;
-            String password = null;
-            WebService.Type st = ServiceChecker.checkService(
-                    DEMO_URL, getBase64EncAuth(userName, password));
-            switch (st) {
-                case Atom:
-                    log.info("Atom Service Found");
-                    break;
-                case WFSOne:
-                    log.info("WFSOne Service Found");
-                    break;
-                case WFSTwo:
-                    log.info("WFSTwo Service Found");
-                    WFSTwoServiceHandler wfstwo =
-                            new WFSTwoServiceHandler(DEMO_URL, userName,
-                                    password);
-                    break;
-                default:
-                    log.log(Level.SEVERE, "Could not determine Service Type");
-                    break;
-            }
-        } else {
-            // Its kind of complicated to start a javafx application from
-            // another class. Thank god for StackExchange:
-            // http://stackoverflow.com/a/25909862
-            new Thread() {
-                @Override
-                public void run() {
-                    javafx.application.Application.launch(Start.class);
-                }
-            }.start();
-            Start start = Start.waitForStart();
-            //start.printInvoking();
+
+        if (help(args)) {
+            helpAndExit();
         }
+
+        if (runHeadless(args)) {
+            Headless.main(args);
+            System.exit(0);
+        }
+
+        // Its kind of complicated to start a javafx application from
+        // another class. Thank god for StackExchange:
+        // http://stackoverflow.com/a/25909862
+        new Thread() {
+            @Override
+            public void run() {
+                javafx.application.Application.launch(Start.class);
+            }
+        }.start();
+        Start start = Start.waitForStart();
+        //start.printInvoking();
     }
 }
